# src/main/java/com/bogdan/releasetracking/dto/UpdateReleaseWsDTO.java
package com.bogdan.releasetracking.dto;

import java.time.LocalDate;

public class UpdateReleaseWsDTO {

    private String name;

    private String description;

    private String status;

    private LocalDate releaseDate;

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }

    public String getStatus() {
        return status;
    }

    public void setStatus(String status) {
        this.status = status;
    }

    public LocalDate getReleaseDate() {
        return releaseDate;
    }

    public void setReleaseDate(LocalDate releaseDate) {
        this.releaseDate = releaseDate;
    }
}